Fixes sample configuration for DotNet
Previously, the sample configuration scripts (`sampleenv.ps1` and `sampleenv.sh`) would set the configuration values to environment variables.

The `Main.cs` shared file previously expected the configuration script to store a configuration file to the users home directory.

This changeset fixes the configuration scripts to write to a file, and changes the expected location to a file that is local to the VS solution.